✅ Split up test helpers
